CTCP-282: Tidy up
