+++ src/test/scala/uk/gov/hmrc/test/ui/pages/TraderDetails/Guarantee/OtherGuarantee3ReferencePage.scala
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pages.TraderDetails.Guarantee
 
-object WantToAddReferenceForGuaranteePage extends BasePage {
+import uk.gov.hmrc.test.ui.pages.BasePage
 
-  val doYouWantToAddReferenceTitle =
-    "Do you want to add a reference for the guarantee? - Manage your transit movements - GOV.UK"
+object OtherGuarantee3ReferencePage extends BasePage {
+
+  val otherGuarantee3ReferenceTitle =
+    "What is the reference for the guarantee? - Manage your transit movements - GOV.UK"
 
   def loadPage: this.type = {
-    onPage(doYouWantToAddReferenceTitle)
+    onPage(otherGuarantee3ReferenceTitle)
     this
   }
 
-  def selectToAddReference(answer: String): WantToAddReferenceForGuaranteePage.type = {
-    answer match {
-      case "Yes" => clickById("value");
-      case "No"  => clickById("value-no");
-    }
+  def enterOtherReference(answer: String): OtherGuarantee3ReferencePage.type = {
+    fillInputById("value", answer)
     this
   }
+
 }
